second part

+++ module14/Final_Project/Compress_Helper.h
@@ -109,7 +109,7 @@ void readBMPFILE(int &width, int &height, int** &image)
                // the Image is a
                // 24-bit BMP Image
                temp = temp & 0x0000FF;
-               image[i][j] = temp;
+               image[i][j] = static_cast<int>(temp);
             }
          }
       }
